feat: Migrate GUI from CustomTkinter to PySide6
This commit migrates the entire graphical user interface of the Shopify Fulfillment Tool from CustomTkinter to PySide6. This migration was undertaken to provide a more robust and scalable foundation for future feature development, as requested by the user.

Key changes include:
- A new main application window (`gui/main_window_pyside.py`) built with PySide6 widgets and layouts.
- A new `PandasModel` (`gui/pandas_model.py`) to efficiently display pandas DataFrames in a `QTableView`, following Qt best practices.
- A new settings window (`gui/settings_window_pyside.py`) with the "General & Paths" tab fully implemented and stubs for more complex configuration tabs.
- The original CustomTkinter entry point has been preserved at `gui_main_tkinter.py`.
- A new main entry point `gui_main.py` now launches the PySide6 application.
- All core business logic from the `shopify_tool/` package remains untouched and is now called by the new GUI.
- The existing test suite has been updated to accommodate the entry point changes, and all 86 tests pass, ensuring no regressions in the backend logic.
