+++ requirements.txt
@@ -9,6 +9,7 @@ openpyxl==3.1.5
 packaging==25.0
 pandas==2.3.1
 pluggy==1.6.0
+PySide6
 pyee==13.0.0
 Pygments==2.19.2
 pyinstaller==6.15.0
